Change int to size_t in Trie impl

+++ src/algo/Trie.cpp
@@ -62,7 +62,7 @@ void Trie::findAllWords(std::shared_ptr<TrieNode> node, const std::string &prefi
         results.push_back(prefix);
     }
 
-    for (int i = 0; i < MAX_CHAR_NO; ++i)
+    for (size_t i = 0; i < MAX_CHAR_NO; ++i)
     {
         if (node->children[i] != nullptr)
         {
